refactored Thing to DbToEs and moved makeRecords there as well
For es bulk upload the slowest part is the json serialization
in pypy3 using the native json is by far the fastests, but
interstingly this is one of those cases where cpython is quite a bit
faster because of rapidjson which brings the serialization time down
to 2 seconds. pypy3 has some major friction around the c api and
rapidjson takes 24 seconds which is even slower than simplejson.