IBNode first pass at handling symmetric predicates
Also added a test for symmetry.

It does seem like it might be reasonable to create
subclasses on the fly that can deal with the symmetry semantics of a
language (like owl2) based on the bound metadata.

Another case that we may need to deal with are order semantics on
predicates such as owl:propertyChainAxiom which induce orderdness on
constructs which are technically unordered.