alt use Accept header, ilxtype + tests
Accept is the correct header to check for the response type not
Content-Type.

fde was missing, so added a test against all the known interlex types
and included the internal interlex type in the output in addition to the
mapped owl type. Also added tests to make sure we have coverage of all
those types.