Queries.reference_host (duh), linked -> connected
reference_host should always have been derived from Queries not from
endpoints, this is now fixed (it only took me 2 months to figure it out).

When referring to subgraphs it makes way more sense to use the adjective
connected, since that is the proper technical term for the relationship
between those subgraphs and the entity they are ultimatley connected to.

Notes about the right way to do qualifiers -- just use the git model with
a few additional relations beyond include and exclude, KISS. Updated
explaning.org with a table providing a rough diagram of the following.
Wow ... the racket version way more readable and easier to understand >_<.