uri structure update for ontologies, update alt to reflect
The ast nodes for generating the uri tree now have their internal node
names separated from the names that are used in the route so that
users don't have to suffer from our implementation details.

The key realization for the ontologies section is that we can just use
the /uris/ convetion again as a sandbox for ontology names while
providing for the version convention as a static get only set of
endpoints.

I am not really satisfied with the implementation, in theory it should
be possible to have the Ontologies class directly implement ilx, uris,
etc. except that it causes some issues, and we would need one more
version of Own and Diff to do the implementation. On the other hand,
it is much easier to detect that you are looking at ontology uris not
regular uris, which is probably the right decision.