test_flows, dbstuff, orcid halfth pass, email ver as stored proc
test_flows has most of what we need to start testing invariants around
permissions etc. specifying the scenarios is a bit trickier, but in
the process able to test various flows by properly walking the db
through real operations to arrive at the starting state for a request
to be send

dbstuff holds sql for mostly login/group/admin stuff which doesn't
make sense in dump/load

the prototype new user flow was wrong because it didn't follow orcid
best practices, which are legitimately vastly better because we don't
have to wonder whether the orcid is valid, we always get it directly
from orcid so we don't have to worry about expiring it etc.

add some api key invariants as a stored proc

run_uri and server_uri add db_kwargs argument ... not sure how i
managed to work without that for so long, pulling stuff from configs
and the environment is a mess for testing when you need to
programatically create a bunch of separate databases