ontspec first pass creation and rudamentary generation done
we should be able to port termsets in as specs during sync now that
the ilx_*/spec routes have been added, they will only accept post for
things that were historically termsets, all termsets should eventually
be mapped to ontology uris, but the old behaior for the historical
termsets will be retained (no now ilx ids will be used for termsets in
the future)

new ingest_ontspec proc that can probably serve as the basis for other
internal ingest use cases where we e.g. have the history of committees
that we want to track via the triples table + identities

updated getGraphByName and friends to support cases where names map
directly to gclc instead of serialization

relaxed restriction on names table because it is cenrtal for mapping
identities, so if anything is uploaded, even an internal interlex
uri (i.e. a spec) we still want it in there to simplify retrieval

in order for the spec routes to work correctly need the fix provided
in https://github.com/pallets/werkzeug/pull/3018

first pass at /ontologies/dns/ routes and reserve the dns group on the
thought that that might serve as an interlex perspective over all
external iris

test_post_ontspec is a good example of how to write a test for post
that allows for rapid development of the functionality needed for a
post endpoint, it is missing validation that the outputs match
expectations, but for hitting the codepaths that need to be developed
it does the trick