wire up logout, first (bad) draft of /<group> route
add note that we do need a set of endpoints to list the types of
things that we have, i think we can probably make it work with
/<group>/types to list the known types and whether they are record
types or resource types
/<group>/types/<rdf-type-curie> then lists all the instance of that,
within reason, e.g. owl:Class def paginated, rdf:Property,
ilxr:Property, ilxr:Relationship, ilxr:Annotation, or something that,
so that it is possible to get e.g. the list of valid values for
e.g. the predicate position (modulo lexical)