sql triples transitive query helper functions
called connected_predicates_{ilx,ext,all} interally for now

as it turns out the performance of filtering triples and views by
predicate and creating a union that can then be queried again is
actually fast enough that the materialized view approach seems
workable, it is a bit more complexity on the query side but avoids
having to modify the triples constraints for a marginal use case,

below are notes from a trigger function variant of the mat view
approach which includes the explaination of why it is very much not
optimal including examples where things will break down, in summary,
if we want to materialize more complex lifting then we will almost
certainly want to do it as part of a process where we produce new
artifacts with the appropriate metadata and prov tracking the lifting
but that would be part of future work related to advanced versioning

CREATE OR REPLACE FUNCTION fun_lift_sco_res_onp_svf()
RETURNS trigger AS $$ DECLARE lift_datatype uri; BEGIN
lift_datatype := 'http://uri.interlex.org/tgbugs/uris/readable/lifts/sco/res-onp-svf';
...
END;

so while in principle this could work, using datatype with s p o but o_lit null would
require extensive modifications to the constraints on the triples table among other things
XXX actually could handle this by having a special case where triple_identity and subgraph_identity are both present
that would minimize the changes needed to the triples table i think?
s p o <->
s sco 0
      0 a res
      0 onp p
      0 svf o

this case illustrates the problem with using subgraph_identity
s p' 0
     0 first o1
     0 rest 1
            1 first on
            1 rest nil
s p' (o1 on)
s eqv 0
      0 int 1
            1 first 3
                    3 a res
                    3 onp p
                    3 svf o1
            1 rest 2
                   2 first 4
                           4 a res
                           4 onp p
                           4 svf on
                   2 rest nil
could also have a rule for which i think i use in neurdf already?
s p' ([p1 o1] [pn on])
anything more complex we won't be able to lift and really
the only things that are worth lifting are the ones that
can be converted to s p o anyway because any others will be
complex enough that querying over them for the interface will
be out of scope, i.e. that they are not strictly hierarchical

i think the proper solution is to do the materialization out of band
but in the same transaction when possible, and to have a direct
relation from the full lifts to the source that was lifted in the
identity relations table, it is not as ... elegant as having the
database do it itself, but it means that we can quickly filter out
lifted sources using the identities table, we can also use the
datatype column to mark lifted triples because lifted triples
should always be s p o and never contain o_lit there, is still the
issue of how to map back to the original subgraph replica and
deduplication that they came from, which works for s p o cases
not for any other