first pass at {group}/query/transitive
also added skeleton for top level dns (for resources that are not
ontologies)

fixed ontologies triple render so that it is simple and not a redirect

the response for query transitive is a graph of only the predicate in
question, the types of the subject and object and their rdfs:label,
the accept header can be used as with any other rdf response

/{group}/query/transitive/{start}/{predicate}?obj-to-sub={}&depth={}
is the full functionality at the moment, obj-to-sub causes queries to
run in the opposite direction which is useful when searching for
children for e.g. the partOf graph, there are many ways that one might
try to name this, but object to subject seems to be understandable and
accurate, scigraph  incomiong/outgoing is confusing

depth currently defaults to unlimited, however when traversing the
expanded graph, e.g. including uberon and all the phenotype ontologies
some other solution will likely be needed due to the expansion to
hundreds of thousands of terms for e.g. anatomical entity (issues with
naming of phenotypes aside) so we will need a cut point or will need
to limit to the source ontology namespace or being defined in the same
ontology file or something like that, or we need to resolve the
multiple traversal issue by just returning all triples with matching
predicates which is fast and then using a better algorithm that won't
traverse subgraphs multiple times